glitchbot garbage removal

+++ Bot Farm Capstone/src/GlitchProjectile.java
@@ -1,7 +1,7 @@
 import processing.core.PImage;
 
 public class GlitchProjectile extends Projectile {
-	public GlitchProjectile(PImage img, int x, int y, int w, int h, String owner, double angle, int delay) {
+	public GlitchProjectile(PImage img, double x, double y, int w, int h, String owner, double angle, int delay) {
         super(img, x, y, w, h, owner, angle, delay);           
 }
 }
